requirements.txt
Django==1.7
Pillow==2.6.1
requests==2.4.3
wsgiref==0.1.2
